Compare job company_id as integers in authorizeJob
Guards against a strict-type mismatch between the job's company_id
and the authenticated company's id (e.g. int vs numeric string from
different query paths), which could wrongly deny access to a job
that legitimately belongs to the company.

+++ app/Http/Controllers/CompanyController.php
@@ -242,7 +242,7 @@ class CompanyController extends Controller
 
     private function authorizeJob(Job $job): void
     {
-        if ($job->company_id !== Auth::user()->company->id) {
+        if ((int) $job->company_id !== (int) Auth::user()->company->id) {
             abort(403);
         }
     }
